backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0

# NLP & ML
spacy==3.8.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
transformers==4.35.2
torch==2.3.1

# Document Processing
pdfplumber==0.10.3
PyMuPDF==1.23.8
python-docx==1.1.0

# RAG & LLM
langchain==0.0.335
chromadb==0.4.18
faiss-cpu==1.10.0
openai==1.3.5
anthropic==0.7.0

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0
lxml==4.9.3

# Utilities
pandas==2.1.3
numpy==1.26.2
regex==2023.10.3
